Shrink channel reordering tables to 9-10 bytes (from 36-40 bytes)

+++ include/vlc_aout.h
@@ -179,8 +179,8 @@ static const uint32_t pi_vlc_chan_order_wg4[] =
  * internal (WG4) order is requested.
  */
 VLC_API unsigned aout_CheckChannelReorder( const uint32_t *, const uint32_t *,
-                                           uint32_t mask, int *table );
-VLC_API void aout_ChannelReorder( void *, size_t, unsigned, const int *, unsigned );
+                                           uint32_t mask, uint8_t *table );
+VLC_API void aout_ChannelReorder( void *, size_t, unsigned, const uint8_t *, unsigned );
 
 /**
  * This fonction will compute the extraction parameter into pi_selection to go
